midterm/requirements.txt
Django	3.1.7
Jinja2	2.11.3
MarkupSafe	1.1.1
PyJWT	1.7.1
pip	21.0.1
psycopg2	2.8.6
sqlparse	0.4.1
Django~=3.1.7
djangorestframework~=3.12.2
abstract~=2020.7.26
djangorestframework-jwt~=1.11.0